mvn-scalaxb: Improve robustness of maven plugin.
The maven plugin now invokes the 'start' method in the Main class
instead of the 'main' method. This allows it to catch exceptions
thrown by 'start' that were otherwise being caught by 'main'.
These are then rethrown as MojoExecutionException (for unexpected
exceptions), or as MojoFailureException (for expected exceptions).

It also avoids calling the scalaxb compiler at all if no XSD files
are found.