fixes xs:group containing optional substitution group generating bad code. fixes #197
a named group generates a parser for mixed complex type,
and the actual bug is in generating parser for mixed.
now the substitution groups are parsed as if they were choices.

ignoreSubGroup parameter was added to buildArgForMixed
to deal distinguish sub group head element
appearing as sub group or as an item of the sub group.