Enumerate all transitive self types in generated SOAP bindings
Problem
-------

In Scala 3 the self type must enumerate all transitive self types. That
means that generated code like this is rejected by the Scala 3
compiler:
```scala
trait HttpClients {}
trait Soap11Clients { this: HttpClients => }
trait FooSoapBindings { this: Soap11Clients => }
```

It complains with:
```
[error] -- [E058] Type Mismatch Error: Test.scala:6:8
[error] 6 |  trait FooSoapBindings { this: Soap11Clients => } // error
[error]   |        ^
[error]   |missing requirement: self type Soap11Clients & FooSoapBindings of trait FooSoapBindings does not conform to self type HttpClients
[error]   |of required trait Soap11Clients
```

Solution
--------

The solution is to add `HttpClients` (or `scalaxb.HttpClientsAsync`) to
the self type of the generated SOAP bindings. Scala 2 is still happy
with this change and Scala 3 is now able to compile the generated
bindings.